Go typescript, add eslint
