+++ packages/tools/remark-plugins/README.md
@@ -16,16 +16,18 @@
 
 ## Format Markdown file
 
+Make sure the package is built (`rush build`).
+
 Format a single Markdown file and output the result to `stdout`:
 
 ```bash
-pnpm dlx remark-cli --use ./packages/tools/remark-plugins/index.js README.md
+pnpm dlx remark-cli --use ./packages/tools/remark-plugins/lib/index.js README.md
 ```
 
 Format a single Markdown file and overwrite the same file:
 
 ```bash
-pnpm dlx remark-cli --use ./packages/tools/remark-plugins/index.js README.md -o
+pnpm dlx remark-cli --use ./packages/tools/remark-plugins/lib/index.js README.md -o
 ```
 
 Use globs (e.g. `*.md`) to format multiple files at once.
