+++ packages/tools/remark-plugins/src/commentMarkers/packageTable.ts
@@ -1,20 +1,20 @@
+import type { RushConfig } from './rush.json';
+
+import { Link, Table, TableRow } from 'mdast';
 import { readFileSync } from 'node:fs';
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import stripJsonComments from 'strip-json-comments';
 
-/**
- * @typedef { import("./rush.json.d.ts").RushConfig } RushConfig
- */
-
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
-const rushConfigPath = join(__dirname, '../../../../rush.json');
+
+const rushConfigPath = join(__dirname, '../../../../../rush.json');
 
 const repoUrl = 'https://github.com/kadena-community/kadena.js';
 const badgeBase = 'https://img.shields.io/npm/v';
 
-const headerRow = {
+const headerRow: TableRow = {
   type: 'tableRow',
   children: [
     {
@@ -28,26 +28,21 @@ const headerRow = {
   ],
 };
 
-/**
- *
- * @returns import('mdast').Root
- */
-export function packageTable() {
+export function packageTable(): Table {
   const contents = readFileSync(rushConfigPath, 'utf-8');
-  /** @type RushConfig */
-  const rushConfig = JSON.parse(stripJsonComments(contents));
+  const rushConfig: RushConfig = JSON.parse(stripJsonComments(contents));
   const projects = rushConfig.projects
     .filter((project) => project.shouldPublish)
     .sort((a, b) => b.packageName.localeCompare(a.packageName));
 
-  const rows = projects.map((project) => {
-    const packageLink = {
+  const rows: TableRow[] = projects.map((project) => {
+    const packageLink: Link = {
       type: 'link',
       url: `${repoUrl}/tree/master/${project.projectFolder}`,
       children: [{ type: 'text', value: project.packageName }],
     };
 
-    const changelogLink = {
+    const changelogLink: Link = {
       type: 'link',
       url: `./${project.projectFolder}/CHANGELOG.md`,
       children: [
