+++ packages/tools/remark-plugins/tsconfig.json
@@ -1,15 +1,11 @@
 {
   "compilerOptions": {
-    "types": ["node"],
-    "allowJs": true,
-    "checkJs": true,
     "forceConsistentCasingInFileNames": true,
-    "lib": ["es2020"],
-    "module": "node16",
+    "module": "esnext",
+    "moduleResolution": "node",
+    "outDir": "lib",
     "skipLibCheck": true,
     "strict": true,
-    "target": "es2020"
-  },
-  "include": ["**/*.js"],
-  "exclude": ["node_modules/"]
+    "types": ["node"]
+  }
 }
